+++ packages/code-du-travail-frontend/package.json
@@ -12,8 +12,8 @@
     "precommit": "lint-staged",
     "start": "next start",
     "tsup": "tsup",
-    "build": "pnpm tsup && node dist/prebuild.js && react-dsfr update-icons && panda codegen && next build",
-    "build:analyze": "pnpm tsup && node dist/prebuild.js && react-dsfr update-icons && panda codegen && ANALYZE=true next build",
+    "build": "pnpm tsup && node dist/prebuild-cli.js && react-dsfr update-icons && panda codegen && next build",
+    "build:analyze": "pnpm tsup && node dist/prebuild-cli.js && react-dsfr update-icons && panda codegen && ANALYZE=true next build",
     "test": "pnpm tsup && ELASTICSEARCH_URL=http://localhost:9200 NEXT_PUBLIC_ES_INDEX_PREFIX=cdtn_test NEXT_PUBLIC_APP_ENV=test NODE_ENV=test node dist/generate-data.js && ELASTICSEARCH_URL=http://localhost:9200 NEXT_PUBLIC_ES_INDEX_PREFIX=cdtn_test NEXT_PUBLIC_APP_ENV=test NODE_ENV=test jest --silent",
     "test:watch": "pnpm test --watch",
     "test:frontend": "NEXT_PUBLIC_APP_ENV=test NODE_ENV=test jest --silent --testPathIgnorePatterns='.*\\.es\\.test\\.ts$|.*\\.script\\.test\\.ts$'",
