+++ packages/code-du-travail-frontend/scripts/prebuild.ts
@@ -50,23 +50,10 @@ export const generateWidgetIntegrity = () => {
   fs.writeFileSync(
     widgetIntegrityOutputPath,
     [
-      "// This file is generated at build time by scripts/prebuild.ts",
+      "// This file is generated at build time by scripts/prebuild-cli.ts",
       "",
       'export const WIDGET_LOADER_INTEGRITY = "' + integrity + '" as const;',
       "",
     ].join("\n")
   );
 };
-
-const run = () => {
-  const isProduction = !!process.env.NEXT_PUBLIC_IS_PRODUCTION_DEPLOYMENT;
-  const host = process.env.NEXT_PUBLIC_SITE_URL ?? "http://localhost:3000";
-  generateRobotsTxt(isProduction, host);
-  console.log("Robots.txt generated.");
-  generateWidgetScript(host);
-  console.log("widget.js generated.");
-  generateWidgetIntegrity();
-  console.log("widgetIntegrity.ts generated.");
-};
-
-run();
