+++ packages/code-du-travail-frontend/src/modules/integration/widgetIntegrity.ts
@@ -1,3 +1,4 @@
-// This file is generated at build time by scripts/prebuild.ts
+// This file is generated at build time by scripts/prebuild-cli.ts
 
-export const WIDGET_LOADER_INTEGRITY = "sha384-GSEnlAvltlfCZ/5CP94DUwe0FLoBfXRnju/gZ9iMpJRAbnZUj9yZp441sl1NtCui" as const;
+export const WIDGET_LOADER_INTEGRITY =
+  "sha384-GSEnlAvltlfCZ/5CP94DUwe0FLoBfXRnju/gZ9iMpJRAbnZUj9yZp441sl1NtCui" as const;
